### 2019-08-29 #1691 @whedon
[PRE REVIEW]: MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna
**Submitting author:** @wingtorres (<a href="http://orcid.org/0000-0002-1411-993X">Walter Torres</a>)
**Repository:** <a href="https://github.com/wingtorres/morphometrix" target ="_blank">https://github.com/wingtorres/morphometrix</a>
**Version:** v1.0.0
