### 2019-08-29 #1691 @labarba
You will also need to add DOIs to your references.
