+++ paper/paper.bib
@@ -1,8 +1,6 @@
-%% Created using Papers on Mon, 26 Aug 2019.
-%% http://papersapp.com/papers/
 
 @article{Ichii:1991,
-author = {Ichii, Taro and Kato, Hidehiro},
+author = {{Ichii}, T. and {Kato}, H.},
 title = {{Food and daily food consumption of southern minke whales in the Antarctic}},
 journal = {Polar Biology},
 year = {1991},
@@ -10,19 +8,23 @@ volume = {11},
 number = {7},
 pages = {479--487},
 month = dec
+doi =  {https://doi.org/10.1007/BF00233083}, 
 }
 
 @article{Clapham:2018,
-author = {Clapham, Phillip J and Ivashchenko, Yulia V},
+author = {Clapham}, P.~J. and {Ivashchenko}, Y.~V.},
 title = {{Whaling catch data are not reliable for analyses of body size shifts}},
 journal = {Nature ecology {\&} evolution},
 year = {2018},
-pages = {1--1},
+volume= {2}
+number = {5}
+pages = {756},
 month = apr
+doi = {https://doi.org/10.1038/s41559-018-0534-2}
 }
 
 @article{French:2011,
-author = {French, Susannah S and Gonz{\'a}lez-Su{\'a}rez, Manuela and Young, Julie K and Durham, Susan and Gerber, Leah R},
+author = {{French}, S.~S. and {Gonz{\'a}lez-Su{\'a}rez}, M. and {Young}, J.~K. and {Durham}, S. and [Gerber], L.~R.},
 title = {{Human Disturbance Influences Reproductive Success and Growth Rate in California Sea Lions (Zalophus californianus)}},
 journal = {PLoS One},
 year = {2011},
@@ -30,17 +32,24 @@ volume = {6},
 number = {3},
 pages = {e17686},
 month = mar
+doi = {https://doi.org/10.1371/journal.pone.0017686}
 }
 
 @article{Clapham:2003,
-author = {Clapham, P J and Berggren, P and Bioscience, S Childerhouse and {2003}},
+author = {{Clapham}, P.~J. and {Berggren}, P. and {Childerhouse}, S. and {Friday}, N.~A. {Kasuya}, T. and {Kell}, L. and {Kock}, K.H.
+          and {Manzanilla-Naim}, S. and {Notabartolo Di Sciara}, G. and {Perrin}, W.~F. and {Read}, A.~J. and {Reeves}, R.~R. and {Rogan}, E.
+          and {Rojas-Bracho}, L. and {Smith}, T.~D. and {Stachowitsch}, M. and {Taylor}, B.~L. and {Thiele}, D. and {Wade}, P.~R and {Brownell}, R.~L.}
 title = {{Whaling as science}},
-journal = {academic.oup.com
-}
+year = {2003},
+journal = {Bioscience},
+volume = {53},
+number = {3},
+pages = {210--212}
+doi = {https://doi.org/10.1641/0006-3568(2003)053[0210:WAS]2.0.CO;2}
 }
 
 @article{Lockyer:1985,
-author = {Lockyer, C H and McConnell, L C and Waters, T D},
+author = {{Lockyer}, C.~H. and {McConnell}, L.~C. and {Waters}, T.~D.},
 title = {{Body condition in terms of anatomical and biochemical assessment of body fat in North Atlantic fin and sei whales}},
 journal = {Canadian Journal of Zoology},
 year = {1985},
@@ -48,112 +57,125 @@ volume = {63},
 number = {10},
 pages = {2328--2338},
 month = jan
+doi = {https://doi.org/10.1139/z85-345}
 }
 
-@article{Christiansen:2013,
[…]
