### 2019-09-05 2d156a4
Update paper.md
