+++ paper/paper.bib
@@ -72,6 +72,16 @@ month = dec,
 doi = {https://doi.org/10.1111/1365-2435.12200}
 }
 
+@article{Christiansen:2018,
+author={Christiansen, Fredrik and Vivier, Fabien and Charlton, Claire and Ward, Rhianne and Amerson, Alicia and Burnell, Stephen and Bejder, Lars},
+title={Maternal body size and condition determine calf growth rates in southern right whales},
+journal={Marine Ecology Progress Series},
+volume={592},
+pages={267--281},
+year={2018},
+doi = {https://doi.org/10.3354/meps12522}
+}
+
 @article{Hilderbrand:1999,
 author = {{Hilderbrand}, G.~V. and {Schwartz}, C.~C. and {Robbins}, C.~T. and {Jacoby}, M.~E. and {Hanley}, T.~A. and {Arthur}, S.~M. and {Servheen}, C.},
 title = {{The importance of meat, particularly salmon, to body size, population productivity, and conservation of North American brown bears}},
