+++ paper/paper.bib
@@ -8,7 +8,7 @@ volume = {11},
 number = {7},
 pages = {479--487},
 month = dec
-doi =  {https://doi.org/10.1007/BF00233083}, 
+doi =  {https://doi.org/10.1007/BF00233083} 
 }
 
 @article{Clapham:2018,
