+++ paper/paper.bib
@@ -7,7 +7,7 @@ year = {1991},
 volume = {11},
 number = {7},
 pages = {479--487},
-month = dec
+month = dec,
 doi =  {https://doi.org/10.1007/BF00233083} 
 }
 
@@ -19,7 +19,7 @@ year = {2018},
 volume= {2}
 number = {5}
 pages = {756},
-month = apr
+month = apr,
 doi = {https://doi.org/10.1038/s41559-018-0534-2}
 }
 
@@ -56,7 +56,7 @@ year = {1985},
 volume = {63},
 number = {10},
 pages = {2328--2338},
-month = jan
+month = jan,
 doi = {https://doi.org/10.1139/z85-345}
 }
 
@@ -168,7 +168,7 @@ journal = {Marine Mammal Science},
 year = {2018},
 volume = {35},
 number = {1},
-pages = {108--139}
+pages = {108--139},
 doi = {https://doi.org/10.1111/mms.12527}
 }
 
@@ -192,7 +192,7 @@ year = {2002},
 volume = {4},
 number = {2},
 pages = {155--164},
-month = oct
+month = oct,
 }
 
 @article{Baker:2000,
