+++ paper/paper.bib
@@ -12,7 +12,7 @@ doi =  {https://doi.org/10.1007/BF00233083}
 }
 
 @article{Clapham:2018,
-author = {Clapham}, P.~J. and {Ivashchenko}, Y.~V.},
+author = {{Clapham}, P.~J. and {Ivashchenko}, Y.~V.},
 title = {{Whaling catch data are not reliable for analyses of body size shifts}},
 journal = {Nature ecology {\&} evolution},
 year = {2018},
