+++ paper/paper.bib
@@ -16,8 +16,8 @@ author = {{Clapham}, P.~J. and {Ivashchenko}, Y.~V.},
 title = {{Whaling catch data are not reliable for analyses of body size shifts}},
 journal = {Nature ecology {\&} evolution},
 year = {2018},
-volume= {2}
-number = {5}
+volume= {2},
+number = {5},
 pages = {756},
 month = apr,
 doi = {https://doi.org/10.1038/s41559-018-0534-2}
@@ -31,7 +31,7 @@ year = {2011},
 volume = {6},
 number = {3},
 pages = {e17686},
-month = mar
+month = mar,
 doi = {https://doi.org/10.1371/journal.pone.0017686}
 }
 
@@ -44,7 +44,7 @@ year = {2003},
 journal = {Bioscience},
 volume = {53},
 number = {3},
-pages = {210--212}
+pages = {210--212},
 doi = {https://doi.org/10.1641/0006-3568(2003)053[0210:WAS]2.0.CO;2}
 }
 
