+++ paper/paper.bib
@@ -38,7 +38,7 @@ doi = {https://doi.org/10.1371/journal.pone.0017686}
 @article{Clapham:2003,
 author = {{Clapham}, P.~J. and {Berggren}, P. and {Childerhouse}, S. and {Friday}, N.~A. {Kasuya}, T. and {Kell}, L. and {Kock}, K.H.
           and {Manzanilla-Naim}, S. and {Notabartolo Di Sciara}, G. and {Perrin}, W.~F. and {Read}, A.~J. and {Reeves}, R.~R. and {Rogan}, E.
-          and {Rojas-Bracho}, L. and {Smith}, T.~D. and {Stachowitsch}, M. and {Taylor}, B.~L. and {Thiele}, D. and {Wade}, P.~R and {Brownell}, R.~L.}
+          and {Rojas-Bracho}, L. and {Smith}, T.~D. and {Stachowitsch}, M. and {Taylor}, B.~L. and {Thiele}, D. and {Wade}, P.~R and {Brownell}, R.~L.},
 title = {{Whaling as science}},
 year = {2003},
 journal = {Bioscience},
