+++ paper/paper.bib
@@ -118,7 +118,7 @@ doi = {https://doi.org/10.1146/annurev-marine-010318-095323}
 @article{Dawson:2017,
 author = {{Dawson}, S.~M.},
 title = {{Inexpensive Aerial Photogrammetry for Studies of Whales and Large Marine Animals}},
-journal = {Frontier in Marine Science}
+journal = {Frontier in Marine Science},
 year = {2017},
 volume = {4}
 pages = {366},
