+++ paper/paper.bib
@@ -120,7 +120,7 @@ author = {{Dawson}, S.~M.},
 title = {{Inexpensive Aerial Photogrammetry for Studies of Whales and Large Marine Animals}},
 journal = {Frontier in Marine Science},
 year = {2017},
-volume = {4}
+volume = {4},
 pages = {366},
 month = nov,
 doi =  {https://doi.org/10.3389/fmars.2017.00366}
