### 2019-09-17 #1691 @isaacvandor
What's the timeline for this? Out on a cruise, but could potentially review
when I get back.

--Isaac
On Mon, Sep 16, 2019, 17:30 carsen-stringer <notifications@github.com>
wrote:

> Sorry this is not my area of expertise
>
