### 2019-10-03 #1691 @theoniphotopoulou
Hi there,
I'm afraid I am completely swamped at the moment. Sorry I can't help!
Best,
Theoni

On Thu, 3 Oct 2019 at 08:45, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @theoniphotopoulou <https://github.com/theoniphotopoulou> @mattdawkins
> <https://github.com/mattdawkins> @meredithmjackson
> <https://github.com/meredithmjackson> @readicculus
> <https://github.com/readicculus> @jgraving <https://github.com/jgraving>
> @alexander-hamme <https://github.com/alexander-hamme> would you be
> interested in reviewing this software
> <https://github.com/wingtorres/morphometrix> package (a python GUI for
> marine animal size analysis from aerial image data), for the Journal of
> Open Source Software (JOSS)? The streamlined review process
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html> takes
> place on GitHub and focuses largely on evaluating the software and on
> reviewing a very short paper
> <https://github.com/openjournals/joss-papers/blob/joss.01691/joss.01691/10.21105.joss.01691.pdf>
> .
>
