### 2019-10-07 #1691 @Kevin-Mattheus-Moerman
@jdpye @mdsumner would you be interested in reviewing [this software](https://github.com/wingtorres/morphometrix) package (a python GUI for marine animal size analysis from aerial image data), for the Journal of Open Source Software (JOSS)? The streamlined [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses largely on evaluating the software and on reviewing [a very short paper](https://github.com/openjournals/joss-papers/blob/joss.01691/joss.01691/10.21105.joss.01691.pdf).
