### 2019-10-22 #1825 @bw4sz
@patrickcgrey throw me some sample whale images to play with here?
