### 2019-10-22 #1825 @bw4sz
grabbing and testing with some data from https://github.com/patrickcgray/cetacean_photogram
