### 2019-10-22 #1825 @bw4sz
testing with this image looks good.
![DSC00917](https://user-images.githubusercontent.com/1208492/67323789-cad2e980-f4c7-11e9-9332-afde908092cb.JPG)
