### 2019-10-22 wingtorres/morphometrix#2 @bw4sz
More text to orient user on home screen
It's a bit cumbersome to toggle between the paper/README/GUI to keep track of each of the steps. Please consider adding some text, each on hover, on in the sidebar that helps guide the user. I think this would reduce the barrier to entry.
Thanks!
