### 2019-10-22 #1825 @bw4sz
added a small idea https://github.com/wingtorres/morphometrix/issues/2
