### 2019-10-25 #1825 @patrickcgray
Hey Ben,

Late in replying but I'm happy to provide more imagery as needed, should be
plenty in that DataDryad repo as you likely saw!

Cheers,
Patrick

On Tue, Oct 22, 2019 at 5:47 PM Kristen Thyng <notifications@github.com>
wrote:

> What a cool picture to get in my email from github notifications!
>
