### 2019-11-01 #1825 @Kevin-Mattheus-Moerman
@wingtorres can you work on adding community guidelines?

Here is some more information on this:
https://help.github.com/en/github/site-policy/github-community-guidelines

Also here is a link to one of my projects. There I link to contributing guidelines in a `CONTRIBUTING.md` file in the `README.md`, which is an approach I recommend.
https://github.com/gibbonCode/GIBBON

If you imagine your contributors community will grow I also recommend adding a `CODE_OF_CONDUCT.md`file too (e.g. based on this template: https://www.contributor-covenant.org/).
