### 2019-11-01 #1825 @Kevin-Mattheus-Moerman
@wingtorres can you work on the issues raised by the reviewers? https://github.com/wingtorres/morphometrix/issues
