### 2019-12-15 a457f7e
add community guidelines, contributing guidelines, integrate README into GUI
