### 2019-12-15 wingtorres/morphometrix#2 @wingtorres
@bw4sz thanks for the suggestion! I agree that this would increase usability. The most recent update integrates the README as a .html the user can scroll through as a widget below the other options on the opening frame. Do you think that does the trick?

![Screen Shot 2019-12-15 at 10 45 58 AM](https://user-images.githubusercontent.com/28067578/70869359-25cd0e00-1f58-11ea-808d-56ec84315346.jpeg)
