### 2019-12-15 #1825 @wingtorres
@danielskatz Yes finally! I have updated the code to address @bw4sz 's suggestion and in regards to @isaacvandor 's thoughts I have at least pointed to where users can customize their version of the code in lieu of adding a preferences menu to modify the GUI cosmetics (although we may want to do this down the line, I believe the code will be useful to people without it for now). @Kevin-Mattheus-Moerman thank you for your suggestion to include the contribute and code of conduct files - I have added those as well. Thanks y'all! Let me know what else needs to be done.
