### 2019-12-16 #1825 @Kevin-Mattheus-Moerman
@isaacvandor, @bw4sz are you happy with the changes implemented? I see all boxes are ticked. Are you happy to recommend acceptance for this work? Thanks!
