### 2019-12-16 #1825 @Kevin-Mattheus-Moerman
@wingtorres thanks for adding community guidelines. One note on your `CONTRIBUTING.md` file (https://github.com/wingtorres/morphometrix/blob/master/CONTRIBUTING.md). It appears it was copied from my own project's file I referred to. Please review the variable naming section since I changed that to suit my project (MATLAB syntax for mesh data handling). Perhaps yours should instead refer to some Python coding style (perhaps check what other major Python projects use there).
