### 2019-12-16 #1825 @wingtorres
You are totally right that's embarrassing :D, sorry I should have paid closer attention. I'm a bit new to Python, but I try to adhere to the PEP-008 standards so I'll post a link to that.
