### 2019-12-21 98bca6d
Add calculate area tool
