+++ morphometrix/__main__.py
@@ -4,7 +4,7 @@ import sys
 import csv
 import numpy as np
 from PyQt5 import QtGui, QtCore
-from PyQt5.QtWidgets import QMainWindow, QApplication, QGraphicsView, QGraphicsScene, QWidget, QPushButton, QCheckBox, QStatusBar, QLabel, QLineEdit, QPlainTextEdit, QTextEdit, QGridLayout, QFileDialog, QGraphicsLineItem, QGraphicsEllipseItem, QGraphicsItem, QMessageBox, QInputDialog
+from PyQt5.QtWidgets import QMainWindow, QApplication, QGraphicsView, QGraphicsScene, QWidget, QPushButton, QCheckBox, QStatusBar, QLabel, QLineEdit, QPlainTextEdit, QTextEdit, QGridLayout, QFileDialog, QGraphicsLineItem, QGraphicsEllipseItem, QGraphicsPolygonItem, QGraphicsItem, QMessageBox, QInputDialog
 from PyQt5.QtWebEngineWidgets import QWebEngineView
 from scipy.special import comb
 
@@ -47,13 +47,17 @@ def main(args=None):
             self.widthsbutton = QPushButton("Measure Widths", self)
             self.widthsbutton.clicked.connect(self.iw.measure_widths)
 
-            self.custom = QPushButton("Custom Length", self)
-            self.custom.clicked.connect(self.measure_custom)
-            self.lengthnames = []
+            self.custombutton = QPushButton("Custom Length", self)
+            self.custombutton.clicked.connect(self.measure_custom)
+            self.lengthNames = []
+
+            self.areabutton = QPushButton("Custom Area", self)
+            self.areabutton.clicked.connect(self.measure_area)
+            self.areaNames = []
 
             self.anglebutton = QPushButton("Measure Angle", self)
             self.anglebutton.clicked.connect(self.measure_angle)
-            self.anglenames = []
+            self.angleNames = []
 
             self.undobutton = QPushButton("Undo", self)
             self.undobutton.clicked.connect(self.undo)
@@ -71,12 +75,14 @@ def main(args=None):
             self.tb.addWidget(self.export)
             self.tb.addWidget(self.lengthbutton)
             self.tb.addWidget(self.widthsbutton)
+            self.tb.addWidget(self.areabutton)
             self.tb.addWidget(self.anglebutton)
-            self.tb.addWidget(self.custom)
+            self.tb.addWidget(self.custombutton)
             self.tb.addWidget(self.undobutton)
             self.tb.addWidget(self.bezier)
 
         def file_open(self):
+            self.iw.scene.clear()
             self.image_name = QFileDialog.getOpenFileName(self, 'Open File')
             self.iw.pixmap = QtGui.QPixmap(self.image_name[0])
             self.iw.pixmap_fit = self.iw.pixmap.scaled(
@@ -84,14 +90,13 @@ def main(args=None):
                 self.iw.pixmap.height(),
                 QtCore.Qt.KeepAspectRatio,
                 transformMode=QtCore.Qt.SmoothTransformation)
-            self.iw.scene.clear()
             self.iw.scene.addPixmap(self.iw.pixmap_fit)  #add image
             self.iw.setScene(self.iw.scene)
 
             #Adjust window size automatically?
             self.setGeometry(
                 QtCore.QRect(QtCore.QPoint(0, 0),
-                            QtCore.QSize(1000, 1000)))  #change main window size
+                            QtCore.QSize(1000, 2000)))  #change main window size
             self.iw.fitInView(self.iw.scene.sceneRect(), QtCore.Qt.KeepAspectRatio)
             self.iw.scene.update()
             self.statusBar.showMessage(
@@ -101,18 +106,20 @@ def main(args=None):
[…]
