### 2019-12-26 #1825 @Kevin-Mattheus-Moerman
> @isaacvandor, @bw4sz are you happy with the changes implemented? I see all boxes are ticked. Are you happy to recommend acceptance for this work? Thanks!

@isaacvandor, @bw4sz :point_up: Thanks again for your help!
