### 2019-12-27 #1825 @wingtorres
hey @Kevin-Mattheus-Moerman thanks for checking in! KC and I had some time over the holidays to improve the GUI and add one feature (measure area), so we would like to make some minor updates to the paper figures and text before publication. Would it be alright if we had until Tuesday Dec. 31 to push these changes before closing review?
