### 2019-12-27 131df35
Consolidate GUI
