+++ morphometrix/__main__.py
@@ -4,274 +4,390 @@ import sys
 import csv
 import numpy as np
 from PyQt5 import QtGui, QtCore
-from PyQt5.QtWidgets import QMainWindow, QApplication, QGraphicsView, QGraphicsScene, QWidget, QPushButton, QCheckBox, QStatusBar, QLabel, QLineEdit, QPlainTextEdit, QTextEdit, QGridLayout, QFileDialog, QGraphicsLineItem, QGraphicsEllipseItem, QGraphicsPolygonItem, QGraphicsItem, QMessageBox, QInputDialog
+from PyQt5.QtWidgets import QMainWindow, QApplication, QGraphicsView, QGraphicsScene, QWidget, QHBoxLayout, QVBoxLayout, QToolBar, QPushButton, QCheckBox, QStatusBar, QLabel, QLineEdit, QPlainTextEdit, QTextEdit, QGridLayout, QFileDialog, QGraphicsLineItem, QGraphicsEllipseItem, QGraphicsPolygonItem, QGraphicsItem, QMessageBox, QInputDialog, QDockWidget, QSizePolicy, QDesktopWidget
 from PyQt5.QtWebEngineWidgets import QWebEngineView
 from scipy.special import comb
 
 #To-do list (descending priority)
-#   -photo saved wonky?
-#   -mouseover xy position?
-#   -preferences change in options?
-#   -tune bezier curve tension parameter on length with scroll wheel (save for patch)
-#   -combine into one window (save for patch)
-#   -angle lines different color (save for patch)
-#   -widths/lengths being incorrectly exported when multiple measurements? (?)
-#   -whale outline (fusiform), concentric whale (possible update in future)
+#   -combine UI into one window (done)
+#   -scale bar
+#   -show values as measurements being made
+#   -mouseover xy position
+#   -preferences change in options
+#   -tune bezier curve tension parameter (rational bezier) with scroll wheel 
+#   -photo saved not working correctly?
+#   -angle lines different color? 
 #   -arrows w/ heads for angle measurement
+#   -arc between angle lines
+#   -object outline: fusiform
 
-def main(args=None):
+# def main(args=None):
+class Manual(QWidget):
+    def __init__(self, parent=None):  
+        super(Manual, self).__init__()
+        self.manual = QWebEngineView()
+        webpage = QtCore.QUrl('https://wingtorres.github.io/morphometrix/')
+        self.manual.setUrl(webpage)
+        self.grid = QGridLayout()
+        self.grid.addWidget(self.manual,1,0)
+        self.setLayout(self.grid)
 
-    #references:
-    #https://stackoverflow.com/questions/26901540/arc-in-qgraphicsscene/26903599#26903599
-    #https://stackoverflow.com/questions/27109629/how-can-i-resize-the-main-window-depending-on-screen-resolution-using-pyqt
-    class Second(QMainWindow):
-        def __init__(self, parent=None):
-            super(Second, self).__init__(parent)
+class Window(QWidget):
 
-            self.iw = imwin()
-            self.setCentralWidget(self.iw)
-            self.setWindowState(self.windowState() & ~QtCore.Qt.WindowMinimized
-                                | QtCore.Qt.WindowActive)
-            self.activateWindow()
+    def __init__(self, parent=None):  
+        #init methods runs every time, use for core app stuff)
+        super(Window, self).__init__()
+        #self.setWindowTitle("MorphoMetriX")
+        #self.setGeometry(50, 50, 100, 200)  #x,y,width,height
+        #self.setStyleSheet("background-color: rgb(0,0,0)") #change color
+        #self.setStyleSheet("font-color: rgb(0,0,0)") #change color
 
-            self.export = QPushButton("Export Measurements", self)
-            self.export.clicked.connect(self.export_meas)
[…]
