+++ _config.yml
@@ -1,1 +1,12 @@
 theme: jekyll-theme-minimal
+plugins:
+  - jekyll-relative-links
+relative_links:
+  enabled: true
+  collections: true
+include:
+  - CONTRIBUTING.md
+  - README.md
+  - LICENSE.md
+  - CODE_OF_CONDUCT.md
+  - CONTRIBUTING.md
