+++ README.md
@@ -3,76 +3,75 @@
 ## Installation 
 [![Anaconda-Server Badge](https://anaconda.org/wingtorres/morphometrix/badges/version.svg)](https://anaconda.org/wingtorres/morphometrix)
 
-The easiest way to get `MorphoMetriX` is to install it into the [Anaconda](https://www.anaconda.com/distribution/) Python distribution. Once Anaconda has been installed, create and/or activate your preferred environment and in terminal enter
+The easiest way to get `MorphoMetriX` is to install it into the [Anaconda](https://www.anaconda.com/distribution/) Python distribution. In Anaconda create and/or activate your preferred environment and in terminal enter
 
     conda install -c wingtorres morphometrix
 
-## Using MorphoMetriX
-After activating and installing into your environment of choice, open the GUI window with the command
+## Running MorphoMetriX
+After activating and installing into your environment of choice, open the GUI with the command
     
     python -m morphometrix
 
-opening the dialog box below. Enter image details, camera specifications, altitude, number of width segments desired, and any notes (see Figure 1-3 in our [JOSS article](???) for examples). 
+In the input frame on the left, enter image details, camera specifications, altitude, number of width segments desired, and any notes (see Figure 1-3 in our JOSS article for examples).
 
 ![image 1](images/Picture1.png)
 
-When ready, select “Start measuring”. A box appears as below. Select “New Image” to import an image (.jpg) for photogrammetric analysis. 
+ Select “New Image” to import an image (.jpg) for photogrammetric analysis.
 
-![image 2](images/Picture2.png)
+__*Measuring*__
 
-*Make Measurements*
-
-Once an image has been imported, you can begin making measurements by selecting “Measure Total Length”, “Custom Length” or “Measure Angle”. Widths can be measured by selecting “Measure Widths” following a total length or custom length measure. The number of width segments desired can be specified in “# Width Segments” in the dialog box. Examples below.
+Once an image has been imported, you can begin making measurements by selecting “Measure Length”, “Measure Area” or “Measure Angle”. The user can label each measurement with a unique name (i.e. “Total Length”). Widths can be measured by selecting “Measure Widths” following any length measurement. The number of width segments desired can be specified in “# Width Segments” in the input frame. Examples below.
 
 *Quick tips*
 
--	Select “Undo” to re-select a point
--	Scroll to zoom in/out
--	Shift + scroll to pan
--	Single click to place a point
--	Double click to complete a length measurement 
 
-*Total length*
+-	**Zoom in/out** - Scroll
+-	**Pan** - <kbd>⇧ Shift</kbd> + scroll
+-	**Place a point** - Single click
+-	**Complete a length measurement** - Double click
+-	**Replace last point** - Select “Undo” 
 
+__*Measure Length*__ (–)
 
-Select “Measure Total Length”. Decide if you want a “Bezier fit?” applied to the length measurement. This is especially useful if the animal is curved, as it will apply a Bezier curve fit to the points selected in the measurement, where the first and last points are fixed. Make single clicks and drag to add line segments and double click to finish the measurement. 
+Select “Measure Length”. A box will appear for the user to create a unique measurement name, i.e. “Total Length”, “Fluke Span”, etc. Decide if you want a “Bezier fit?” applied to the length measurement. This is especially useful if the animal is curved, as it will apply a smooth fitting curve to the points selected in the measurement. Make single clicks to make a measurement and double click to finish. 
 
-![image 3](images/Picture3.png)
+![image 3](images/Picture2.png)
 
-*Measuring Widths*
+__*Measuring Widths*__ ($+$)
 
[…]
