### 2019-12-31 9c7608a
Update manual
