### 2019-12-31 2528898
Update paper figures and text
