+++ paper/paper.md
@@ -29,28 +29,27 @@ Body size is recognized as one of the most important factors that determine anim
 
 Several software systems, written in MATLAB and R, have been developed to extract photogrammetric measures from aerial images of megavertebrates. However, some of these tools have slow workflows, are designed with only particular aircraft/sensor combinations and/or specific measurements in mind (e.g. [@Dawson:2017; Burnett et al., 2018]), have limited image zoom capabilities (e.g. [@Christiansen:2016; @Dawson:2017; @Burnett:2018]) and in some cases are not open source, requiring purchased or institutional licensing (e.g. [@Dawson:2017; @Burnett:2018]). These existing tools written in R and MATLAB are relatively feature-rich, allowing the measurement of proportional body widths based on total length, as well as a set of additional pre-fixed morphometric measurement options, such as ‘width of eyes’, ‘rostrum to blowhole’, and ‘fluke width’.  However, these predetermined measurements are the only ones supported by the software system and can be altered and/or adapted only if the user has knowledge of the specific coding language, in these cases MATLAB or R.
 
-``MorphoMetriX`` is an open-source application that overcomes some of the limitations of these existing tools, combining the powerful zoom, accuracy, and measurement freedom of ImageJ, the accessibility of a GUI designed to measure body widths based on body length, and the speed of the PyQt5 codebase. It was designed as a simple to use and accurate program for robust morphometric analysis that does not require knowledge of any scripting language for customization. ``MorphoMetriX`` allows the user to input flight and sensor parameters, such as altitude, focal length, and pixel dimensions (calculated by dividing the sensor width of the camera by the image width in pixels), so that all measurements in pixels are automatically scaled to real world values (i.e. meters) (Fig. 1-3). The user imports an image of an animal and can then measure the “Total Length” and “Width” (the user can specify the number of width segments, Fig. 2 & 3), as well as an option for “Custom Length” where the user can label and create custom measurements, such as “Fluke Span”, “Fluke Chord Width”, “Pectoral Fin Width”, “Scar”, etc. (Fig. 3). An option also exists to “Measure Angle” where the user can create a custom angle measurement, such as “Sweep Angle” (Fig. 3). Once completed, all measurements and their labels are exported into a .csv, along with an image of all the measurements that were made on the animal. 
+MorphoMetriX is an open-source application that overcomes some of the limitations of these existing tools, combining the powerful zoom, accuracy, and measurement freedom of ImageJ, the accessibility of a GUI designed to measure body widths based on body length, and the speed of the PyQt5 codebase. It was designed as a simple to use and accurate program for robust morphometric analysis that does not require knowledge of any scripting language for customization. MorphoMetriX allows the user to input flight and sensor parameters, such as altitude, focal length, and pixel dimensions (calculated by dividing the sensor width of the camera by the image width in pixels), so that all measurements in pixels are automatically scaled to real world values (i.e. meters) (Fig. 1-3). The user imports an image of an animal and can create custom length measurements, such as “total length”, “fluke span”, “fluke chord width”, “pectoral fin width”, “scar”, etc (Fig. 2). The user can measure perpendicular widths based on a length measurement (the user can specify the number of width segments, Fig. 2 & 3). Additionally, the user has the option to “Measure Area” , where they can create custom area measurements such as “diatom patch” or “fluke area”, and to “Measure Angle” where the user can create a custom angle measurement, such as “Sweep Angle” (Fig. 2). Once completed, all measurements and their labels are exported into a .csv, along with an image of all the measurements that were made on the animal. 
 
[…]
