### 2020-01-11 #1825 @Kevin-Mattheus-Moerman
@wingtorres I have one minor comment on the paper. The accepted version will have links to your repo so please remove the last sentence:
`Source code and manual are available at https: //github.com/wingtorres/morphometrix/`
