+++ paper/paper.md
@@ -31,7 +31,7 @@ Several software systems, written in MATLAB and R, have been developed to extrac
 
 MorphoMetriX is an open-source application that overcomes some of the limitations of these existing tools, combining the powerful zoom, accuracy, and measurement freedom of ImageJ, the accessibility of a GUI designed to measure body widths based on body length, and the speed of the PyQt5 codebase. It was designed as a simple to use and accurate program for robust morphometric analysis that does not require knowledge of any scripting language for customization. MorphoMetriX allows the user to input flight and sensor parameters, such as altitude, focal length, and pixel dimensions (calculated by dividing the sensor width of the camera by the image width in pixels), so that all measurements in pixels are automatically scaled to real world values (i.e. meters) (Fig. 1-3). The user imports an image of an animal and can create custom length measurements, such as “total length”, “fluke span”, “fluke chord width”, “pectoral fin width”, “scar”, etc (Fig. 2). The user can measure perpendicular widths based on a length measurement (the user can specify the number of width segments, Fig. 2 & 3). Additionally, the user has the option to “Measure Area” , where they can create custom area measurements such as “diatom patch” or “fluke area”, and to “Measure Angle” where the user can create a custom angle measurement, such as “Sweep Angle” (Fig. 2). Once completed, all measurements and their labels are exported into a .csv, along with an image of all the measurements that were made on the animal. 
 
-``MorphoMetriX`` has already been used on several projects with a variety of cetacean species, including bottlenose dolphins, blue, humpback, fin, Antarctic minke, dwarf minke, pilot, and beaked whales. It is also ideal for training and teaching in the classroom for lessons and labs in photogrammetry and morphometrics. Source code and manual are available at https://github.com/wingtorres/morphometrix/
+``MorphoMetriX`` has already been used on several projects with a variety of cetacean species, including bottlenose dolphins, blue, humpback, fin, Antarctic minke, dwarf minke, pilot, and beaked whales. It is also ideal for training and teaching in the classroom for lessons and labs in photogrammetry and morphometrics.
 
 # Figures
 ![Photogrammetry](../images/figure1.png)
