### 2020-01-13 9405d80
minor text edits
