+++ meta.yaml
@@ -2,7 +2,7 @@
 
 package:
   name: morphometrix 
-  version: '1.0.1'
+  version: '1.0.2'
 
 source:
   #path: morphometrix/
