### 2020-01-13 #1825 @wingtorres
- [x] sentence removed
- [x] software archived. DOI is 10.5281/zenodo.3606865 URL to repository is https://zenodo.org/record/3606865#.XhzlQhdKiuU
- [x] released new version (v1.0.2) that is now archived on Zenodo
