### 2020-01-14 #1825 @Kevin-Mattheus-Moerman
@wingtorres
- [ ] Please update the Zenodo meta-data such that the author names and the title match that of the JOSS paper
