### 2020-01-14 #1825 @wingtorres
@Kevin-Mattheus-Moerman
- [x] edited the "basic information" section of the archive to include our names and paper title and published the updated archive on zenodo
