+++ paper/paper.bib
@@ -8,7 +8,7 @@ volume = {11},
 number = {7},
 pages = {479--487},
 month = dec,
-doi =  {https://doi.org/10.1007/BF00233083} 
+doi =  {10.1007/BF00233083} 
 }
 
 @article{Clapham:2018,
@@ -20,7 +20,7 @@ volume= {2},
 number = {5},
 pages = {756},
 month = apr,
-doi = {https://doi.org/10.1038/s41559-018-0534-2}
+doi = {10.1038/s41559-018-0534-2}
 }
 
 @article{French:2011,
@@ -32,7 +32,7 @@ volume = {6},
 number = {3},
 pages = {e17686},
 month = mar,
-doi = {https://doi.org/10.1371/journal.pone.0017686}
+doi = {10.1371/journal.pone.0017686}
 }
 
 @article{Clapham:2003,
@@ -45,7 +45,7 @@ journal = {Bioscience},
 volume = {53},
 number = {3},
 pages = {210--212},
-doi = {https://doi.org/10.1641/0006-3568(2003)053[0210:WAS]2.0.CO;2}
+doi = {10.1641/0006-3568(2003)053[0210:WAS]2.0.CO;2}
 }
 
 @article{Lockyer:1985,
@@ -57,7 +57,7 @@ volume = {63},
 number = {10},
 pages = {2328--2338},
 month = jan,
-doi = {https://doi.org/10.1139/z85-345}
+doi = {10.1139/z85-345}
 }
 
 @article{Christiansen:2014,
@@ -69,7 +69,7 @@ volume = {28},
 number = {3},
 pages = {579--588},
 month = dec,
-doi = {https://doi.org/10.1111/1365-2435.12200}
+doi = {10.1111/1365-2435.12200}
 }
 
 @article{Christiansen:2018,
@@ -79,7 +79,7 @@ journal={Marine Ecology Progress Series},
 volume={592},
 pages={267--281},
 year={2018},
-doi = {https://doi.org/10.3354/meps12522}
+doi = {10.3354/meps12522}
 }
 
 @article{Hilderbrand:1999,
@@ -90,7 +90,7 @@ year = {1999},
 volume = {77},
 number = {1},
 pages = {132--138},
-doi = {https://doi.org/10.1139/z98-195}
+doi = {10.1139/z98-195}
 }
 
 @article{Miller:2012,
@@ -101,7 +101,7 @@ year = {2012},
 volume = {459},
 pages = {135--156},
 month = jul,
-doi = {https://doi.org/10.3354/meps09675}
+doi = {10.3354/meps09675}
 }
 
 @article{Johnston:2019,
@@ -112,7 +112,7 @@ year = {2019},
 volume = {11},
 number = {1},
 pages = {439--463},
-doi = {https://doi.org/10.1146/annurev-marine-010318-095323}
+doi = {10.1146/annurev-marine-010318-095323}
 }
 
 @article{Dawson:2017,
@@ -123,7 +123,7 @@ year = {2017},
 volume = {4},
 pages = {366},
 month = nov,
-doi =  {https://doi.org/10.3389/fmars.2017.00366}
+doi =  {10.3389/fmars.2017.00366}
 }
 
 @article{Goldbogen:2015,
@@ -135,7 +135,7 @@ volume = {93},
 number = {9},
 pages = {687--700},
 month = jul,
-doi = {https://doi.org/10.1139/cjz-2014-0311}
+doi = {10.1139/cjz-2014-0311}
 }
 
 @article{Altmann:1993,
@@ -146,7 +146,7 @@ year = {1993},
 volume = {30},
 number = {2},
 pages = {149--161},
[…]
