### 2020-01-15 #1825 @wingtorres
hey @arfon @kyleniemeyer , I edited the paper.bib doc and removed the 'https://doi.org/' prefixes, but am still getting invalid DOI's after checking references. Do y'all know how to sort this out?
