### 2020-01-15 #1825 @arfon
>  I'm not sure what to do about the invalid DOI though...it should link to the article below. https://academic.oup.com/bioscience/article/53/3/210/251944

I think you can ignore this. The URL library we're using for Whedon doesn't like https://doi.org/10.1641/0006-3568(2003)053[0210:WAS]2.0.CO;2 (it's a pretty unusual URL) but it looks like it resolves.
