### 2020-01-15 #1825 @kyleniemeyer
Hi @wingtorres, looks like the references are fine, but I have just a few typographic issues to resolve:
- Can you remove the period at the end of the title?
- in a number of places, you are missing commas after "i.e." and "e.g."
- in the 4th paragraph, there are some inconsistencies in some citation formatting; you can use these syntax options for in-text citations:
```
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```

Otherwise, it is well written, thanks!
