### 2020-01-15 b5ba28a
typographic edits
