+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna.'
+title: 'MorphoMetriX: a photogrammetric measurement GUI for morphometric analysis of megafauna'
 tags:
   - Python
   - GUI
@@ -23,18 +23,18 @@ bibliography: paper.bib
 
 Body size is recognized as one of the most important factors that determine animal function and performance in its environment [@SchmidtNielsen:1984] and well-sampled and accurate size measurements of wild animals can indicate population health [@Altmann:1993; @Hilderbrand:1999; @French:2011]. However, estimating the body size of large wild vertebrates, like many cetaceans (whales, dolphins, and porpoises) is particularly challenging, as their mobility and large size impedes in situ measurements or live capture [@Huang:2009; @Goldbogen:2015]. Historically, lethal sampling, either scientifically or opportunistically from commercial whaling, has provided measurements of large whales in the past [@Mackintosh:1929; @Lockyer:1985; @Ichii:1991; @Christiansen:2014], but these methods are often not permissible today, expensive, and/or are considered unethical and biased [@Baker:2000; @Clapham:2003; @Clapham:2018]. Alternatively, aerial photogrammetry has proven to be a reliable non-invasive method for obtaining measurements of many megavertebrates, such as cetaceans [@Perryman:2002; @Miller:2012], and the recent advancement of unoccupied aircraft systems (UAS, or drones) has enabled a more affordable, efficient, and accessible means of acquiring high-resolution aerial imagery for morphometric analysis [@Durban:2015; @Christiansen:2018; @Johnston:2019]. With increased opportunity to collect aerial imagery of inaccessible wildlife via UAS, an efficient and simple to use tool for accurate morphometric measurements is needed. 
 
-``MorphoMetriX`` is a flexible photogrammetry graphical user interface (GUI) developed in PyQt5 for making efficient manual morphometric measurements of wild animals via aerial imagery. It was originally designed for the purpose of obtaining morphometric measurements of large whales imaged via UAS, but can be applied to other animals (e.g. pinnipeds, alligators, manatees) and used with nadir imagery collected via other platforms. 
+``MorphoMetriX`` is a flexible photogrammetry graphical user interface (GUI) developed in PyQt5 for making efficient manual morphometric measurements of wild animals via aerial imagery. It was originally designed for the purpose of obtaining morphometric measurements of large whales imaged via UAS, but can be applied to other animals (e.g., pinnipeds, alligators, manatees) and used with nadir imagery collected via other platforms. 
 
-``MorphoMetriX`` was inspired by the image processing program ImageJ [@Schneider:2012] and other UAS photogrammetry tools written in R [@Christiansen:2016] and MATLAB [@Dawson:2017; @Burnett:2018] that were developed specifically for measuring body condition of cetaceans. While ImageJ is an open source software with a powerful zoom function and the freedom and flexibility to measure any object of interest in an image, it lacks the ability to automatically section animals proportionally for perpendicular width measurements (i.e. body width measurements at 10% intervals along the total length of an animal), a key method used for comparative studies of body condition in cetaceans [@Miller:2012; @Christiansen:2016; @Dawson:2017; @Christiansen:2018].
[…]
