### 2020-01-15 #1825 @wingtorres
@kyleniemeyer thank you for the typographic edits! Just made the changes you suggested and generated a new proof.
