### 2020-01-16 #1825 @kyleniemeyer
Congrats on your article's publication in JOSS @wingtorres!
